requirements.txt
openai>=1.0.0
rank-bm25>=0.2.2
numpy>=1.24.0
sentence-transformers>=2.2.0
tqdm>=4.60.0
python-dotenv>=1.0.0
